+++ src/main/java/shop/mtcoding/metamall/core/session/LoginUser.java
@@ -5,11 +5,11 @@ import lombok.Getter;
 
 @Getter
 public class LoginUser {
-    private Integer id;
+    private Long id;
     private String role;
 
     @Builder
-    public LoginUser(Integer id, String role) {
+    public LoginUser(Long id, String role) {
         this.id = id;
         this.role = role;
     }
